+++ frontend/src/components/data-products/data-output-form/data-output-configuration-form.component.tsx
@@ -155,7 +155,6 @@ export function DataOutputConfigurationForm({
                     />
                 );
                 break;
-            case 'string':
             default:
                 inputComponent = <Input />;
                 break;